+++ src/lib/stores/index.ts
@@ -5,3 +5,4 @@ export * from './vault';
 export * from './editor';
 export * from './search';
 export * from './graph';
+export * from './config';